rvc/requirements.txt
torch>=2.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.23.0
scipy>=1.10.0
pyworld>=0.3.0
praat-parselmouth>=0.4.0
faiss-cpu>=1.7.0
